+++ test/postdeploy.test.js
@@ -41,7 +41,7 @@ describe('Running Post-Deployment Integration Tests', () => {
       }).catch((e) => {
         throw e;
       });
-  }).timeout(10000);
+  }).timeout(60000);
 
   it('Spark srcset', async () => {
     await chai
@@ -53,7 +53,7 @@ describe('Running Post-Deployment Integration Tests', () => {
       }).catch((e) => {
         throw e;
       });
-  });
+  }).timeout(10000);
 
   it('Unsplash srcset', async () => {
     await chai
@@ -66,7 +66,7 @@ describe('Running Post-Deployment Integration Tests', () => {
       }).catch((e) => {
         throw e;
       });
-  });
+  }).timeout(10000);
 
   it('Service reports status', async () => {
     await chai
@@ -78,5 +78,72 @@ describe('Running Post-Deployment Integration Tests', () => {
       }).catch((e) => {
         throw e;
       });
-  });
+  }).timeout(10000);
+}).timeout(10000);
+
+describe('Running Post-Deployment Integration Tests on Preprod', () => {
+  it('Youtube OEmbed', async () => {
+    await chai
+      .request('https://preprod.adobeioruntime.net/')
+      .get(`${getbaseurl()}/https://www.youtube.com/watch?v=TTCVn4EByfI`)
+      .then((response) => {
+        expect(response).to.have.status(200);
+        expect(response.text).to.contain('youtube.com');
+        expect(response.text).to.contain('iframe');
+        expect(response.text).to.contain('oembed');
+      }).catch((e) => {
+        throw e;
+      });
+  }).timeout(60000);
+
+  it('Youtube OEmbed (Escaped)', async () => {
+    await chai
+      .request('https://preprod.adobeioruntime.net/')
+      .get(`${getbaseurl()}/https%3A%2F%2Fwww.youtube.com%2Fwatch%3Fv%3DTTCVn4EByfI`)
+      .then((response) => {
+        expect(response).to.have.status(200);
+        expect(response.text).to.contain('youtube.com');
+        expect(response.text).to.contain('iframe');
+        expect(response.text).to.contain('oembed');
+      }).catch((e) => {
+        throw e;
+      });
+  }).timeout(10000);
+
+  it('Spark srcset', async () => {
+    await chai
+      .request('https://preprod.adobeioruntime.net/')
+      .get(`${getbaseurl()}/https://spark.adobe.com/post/z4eHLkF8nZII1/`)
+      .then((response) => {
+        expect(response).to.have.status(200);
+        expect(response.text).to.contain('srcset');
+      }).catch((e) => {
+        throw e;
+      });
+  }).timeout(10000);
+
+  it('Unsplash srcset', async () => {
+    await chai
+      .request('https://preprod.adobeioruntime.net/')
+      .get(`${getbaseurl()}/https://unsplash.com/photos/0lD9SSMC6jo`)
+      .then((response) => {
+        expect(response).to.have.status(200);
+        expect(response.text).to.contain('Unsplash');
+        expect(response.text).to.contain('srcset');
+      }).catch((e) => {
+        throw e;
+      });
+  }).timeout(10000);
+
+  it('Service reports status', async () => {
+    await chai
+      .request('https://preprod.adobeioruntime.net/')
+      .get(`${getbaseurl()}/_status_check/healthcheck.json`)
+      .then((response) => {
+        expect(response).to.have.status(200);
+        expect(response).to.have.header('Content-Type', 'application/json');
+      }).catch((e) => {
+        throw e;
+      });
+  }).timeout(10000);
 }).timeout(10000);